GOV-14601 bulk load optimizations for large promote jobs.

+++ igx.jobs.bulkloadprocessor/Program.cs
@@ -83,20 +83,24 @@ namespace igx.jobs.bulkloadprocessor
                     companyConnection.Close();
 
                     if (loadItemRowCount <= 0)
-                    {
-                        byte[] bytes;
+                    {                        
+                        SLDocument xls = null;
                         if (load.File == null)
-                        {
-                            var data = storage.GetFileContentsAsString($"{constants.COMPANY_BULK_LOAD_FOLDER}",$"{loadInfo.CompanyID}/load_{load.ID}.{load.Extension}");
-                            bytes = Encoding.Default.GetBytes(data);
+                        {        
+                            using (MemoryStream stream = new MemoryStream())
+                            {
+                                await storage.GetFileStream($"{constants.COMPANY_BULK_LOAD_FOLDER}", $"{loadInfo.CompanyID}/load_{load.ID}.{load.Extension}", stream);
+                                
+                                xls = new SLDocument(stream);                                
+                            }                            
                         }
                         else
-                        {
-                            bytes = load.File;
+                        {                            
+                            var memoryStream = new MemoryStream(load.File);
+                            xls = new SLDocument(memoryStream);
                         }
 
-                        var memoryStream = new MemoryStream(bytes);
-                        var xls = new SLDocument(memoryStream);
+                        
 
                         var stats = xls.GetWorksheetStatistics();
 
@@ -249,28 +253,31 @@ namespace igx.jobs.bulkloadprocessor
 
                         #region Update Lookup Values
 
-                        using (var trans = companyConnection.BeginTransaction())
+                        var lookupColumns = loadColumns.Where(l => l.IsLookup);
+
+                        if (lookupColumns.Any())
                         {
-
-                            var lookupColumns = loadColumns.Where(l => l.IsLookup);
-                            List<dynamic> tempLookupColumns = new List<dynamic>();
-
-                            foreach(var col in lookupColumns)
+                            using (var trans = companyConnection.BeginTransaction())
                             {
-                                var loadCol = load.LoadColumns.FirstOrDefault(l => l.Name == col.Name);
+                                List<dynamic> tempLookupColumns = new List<dynamic>();
 
-                                if (loadCol != null && col.FieldTypeId != null)
+                                foreach (var col in lookupColumns)
                                 {
-                                    tempLookupColumns.Add(new {
-                                        LoadID = load.ID,
-                                        loadCol.ColumnIndex,
-                                        col.Name,
-                                        FieldTypeID = col.FieldTypeId
-                                    });
-                                }
-                            }
+                                    var loadCol = load.LoadColumns.FirstOrDefault(l => l.Name == col.Name);
 
-                            companyConnection.Execute(@"drop table if exists #tempLookupColumns;
+                                    if (loadCol != null && col.FieldTypeId != null)
+                                    {
+                                        tempLookupColumns.Add(new
+                                        {
+                                            LoadID = load.ID,
+                                            loadCol.ColumnIndex,
+                                            col.Name,
+                                            FieldTypeID = col.FieldTypeId
+                                        });
+                                    }
+                                }
+
+                                companyConnection.Execute(@"drop table if exists #tempLookupColumns;
                             create table #tempLookupColumns (
                                 LoadID int,
                                 Name nvarchar(max),
@@ -279,46 +286,48 @@ namespace igx.jobs.bulkloadprocessor
                                 );", transaction: trans);
 
 
-                            using (var bulkCopy = new SqlBulkCopy(companyConnection, SqlBulkCopyOptions.Default, trans))
-                            {
 
-                                bulkCopy.BatchSize = SqlBulkBatchSize;
-                                bulkCopy.DestinationTableName = "#tempLookupColumns";
-                                bulkCopy.BulkCopyTimeout = 3600;
-
-                                var table = new System.Data.DataTable();
-                                var columnName = "LoadID";
-                                table.Columns.Add(columnName, typeof(int));
-                                bulkCopy.ColumnMappings.Add(columnName, columnName);
-
-                                columnName = "Name";
-                                table.Columns.Add(columnName, typeof(string));
-                                bulkCopy.ColumnMappings.Add(columnName, columnName);
-
-                                columnName = "ColumnIndex";
-                                table.Columns.Add(columnName, typeof(int));
-                                bulkCopy.ColumnMappings.Add(columnName, columnName);
-
-                                columnName = "FieldTypeID";
-                                table.Columns.Add(columnName, typeof(int));
-                                bulkCopy.ColumnMappings.Add(columnName, columnName);
-
-                                foreach (var item in tempLookupColumns)
+                                using (var bulkCopy = new SqlBulkCopy(companyConnection, SqlBulkCopyOptions.Default, trans))
                                 {
-                                    var row = table.NewRow();
 
-                                    row["LoadID"] = item.LoadID;
-                                    row["Name"] = item.Name;
-                                    row["ColumnIndex"] = item.ColumnIndex;
-                                    row["FieldTypeID"] = item.FieldTypeID;
+                                    bulkCopy.BatchSize = SqlBulkBatchSize;
+                                    bulkCopy.DestinationTableName = "#tempLookupColumns";
+                                    bulkCopy.BulkCopyTimeout = 3600;
 
-                                    table.Rows.Add(row);
+                                    var table = new System.Data.DataTable();
+                                    var columnName = "LoadID";
+                                    table.Columns.Add(columnName, typeof(int));
+                                    bulkCopy.ColumnMappings.Add(columnName, columnName);
+
+                                    columnName = "Name";
+                                    table.Columns.Add(columnName, typeof(string));
+                                    bulkCopy.ColumnMappings.Add(columnName, columnName);
+
+                                    columnName = "ColumnIndex";
+                                    table.Columns.Add(columnName, typeof(int));
+                                    bulkCopy.ColumnMappings.Add(columnName, columnName);
+
+                                    columnName = "FieldTypeID";
+                                    table.Columns.Add(columnName, typeof(int));
+                                    bulkCopy.ColumnMappings.Add(columnName, columnName);
+
+                                    foreach (var item in tempLookupColumns)
+                                    {
+                                        var row = table.NewRow();
+
+                                        row["LoadID"] = item.LoadID;
+                                        row["Name"] = item.Name;
+                                        row["ColumnIndex"] = item.ColumnIndex;
+                                        row["FieldTypeID"] = item.FieldTypeID;
+
+                                        table.Rows.Add(row);
+                                    }
+
+                                    bulkCopy.WriteToServer(table);
                                 }
 
-                                bulkCopy.WriteToServer(table);
-                            }
 
-                            companyConnection.Execute(@"
+                                companyConnection.Execute(@"
                                 declare @maxlen int = 0;
 
                                 drop table if exists #TempBulkLookupValues;
@@ -348,7 +357,8 @@ namespace igx.jobs.bulkloadprocessor
 
                                 ", transaction: trans, commandTimeout: 3600);
 
-                            trans.Commit();
+                                trans.Commit();
+                            }
                         }
 
                         #endregion
